+++ include/z64.h
@@ -101,62 +101,6 @@ typedef enum {
 } KaleidoOverlayType;
 
 typedef struct {
-    /* 0x00 */ GameState state;
-} SetupState; // size = 0xA4
-
-typedef struct {
-    /* 0x0000 */ GameState state;
-    /* 0x00A4 */ u8* staticSegment;
-    /* 0x00A8 */ View view;
-    /* 0x01D0 */ SramContext sramCtx;
-    /* 0x01D4 */ u16 unk_1D4; // not used in mq dbg (some sort of timer that doesn't seem to affect anything)
-    /* 0x01D6 */ s16 coverAlpha;
-    /* 0x01D8 */ s16 addAlpha; // not used in mq dbg
-    /* 0x01DA */ u16 visibleDuration; // not used in mq dbg
-    /* 0x01DC */ s16 ult;
-    /* 0x01DE */ s16 uls;
-    /* 0x01E0 */ char unk_1E0[0x01];
-    /* 0x01E1 */ u8 exit;
-    /* 0x01E2 */ char unk_1E2[0x06];
-} ConsoleLogoState; // size = 0x1E8
-
-struct MapSelectState;
-
-typedef struct {
-    /* 0x00 */ char* name;
-    /* 0x04 */ void (*loadFunc)(struct MapSelectState*, s32);
-    /* 0x08 */ s32 entranceIndex;
-} SceneSelectEntry; // size = 0xC
-
-typedef struct MapSelectState {
-    /* 0x0000 */ GameState state;
-    /* 0x00A8 */ View view;
-    /* 0x01D0 */ s32 count;
-    /* 0x01D4 */ SceneSelectEntry* scenes;
-    /* 0x01D8 */ s32 currentScene;
-    /* 0x01DC */ s32 pageDownIndex; // Index of pageDownStops
-    /* 0x01E0 */ s32 pageDownStops[7];
-    /* 0x01FC */ char unk_1FC[0x0C];
-    /* 0x0208 */ s32 opt;
-    /* 0x020C */ s32 topDisplayedScene; // The scene which is currently at the top of the screen
-    /* 0x0210 */ char unk_210[0x0C];
-    /* 0x021C */ s32 verticalInputAccumulator;
-    /* 0x0220 */ s32 verticalInput;
-    /* 0x0224 */ s32 timerUp;
-    /* 0x0228 */ s32 timerDown;
-    /* 0x022C */ s32 lockUp;
-    /* 0x0230 */ s32 lockDown;
-    /* 0x0234 */ s32 unk_234; // unused
-    /* 0x0238 */ u8* staticSegment;
-} MapSelectState; // size = 0x240
-
-typedef struct {
-    /* 0x0000 */ GameState state;
-    /* 0x00A4 */ u8* staticSegment;
-    /* 0x00A8 */ View view;
-} SampleState; // size = 0x1D0
-
-typedef struct {
     /* 0x00 */ u8 byte0;
     /* 0x01 */ u8 byte1;
     /* 0x02 */ u8 byte2;
@@ -263,90 +207,6 @@ typedef struct PlayState {
     /* 0x1242C */ SceneTableEntry* loadedScene;
     /* 0x12430 */ char unk_12430[0xE8];
 } PlayState; // size = 0x12518
-
-typedef struct {
-    /* 0x0000 */ GameState state;
-    /* 0x00A8 */ View view;
-} TitleSetupState; // size = 0x1D0
-
-typedef struct {
-    /* 0x00000 */ GameState state;
-    /* 0x000A4 */ Vtx* windowVtx;
-    /* 0x000A8 */ u8* staticSegment;
-    /* 0x000AC */ u8* parameterSegment;
-    /* 0x000B0 */ char unk_B0[0x8];
-    /* 0x000B8 */ View view;
-    /* 0x001E0 */ SramContext sramCtx;
-    /* 0x001E4 */ char unk_1E4[0x4];
-    /* 0x001E8 */ SkyboxContext skyboxCtx;
-    /* 0x00348 */ MessageContext msgCtx;
-    /* 0x0E760 */ Font font;
-    /* 0x1C8E8 */ EnvironmentContext envCtx;
-    /* 0x1C9E4 */ char unk_1C9E4[0x4];
-    /* 0x1C9E8 */ Vtx* windowContentVtx;
-    /* 0x1C9EC */ Vtx* keyboardVtx;
-    /* 0x1C9F0 */ Vtx* nameEntryVtx;
-    /* 0x1C9F4 */ u8 n64ddFlag;
-    /* 0x1C9F6 */ u16 deaths[3];
-    /* 0x1C9FC */ u8 fileNames[3][8];
-    /* 0x1CA14 */ u16 healthCapacities[3];
-    /* 0x1CA1C */ u32 questItems[3];
-    /* 0x1CA28 */ s16 n64ddFlags[3];
-    /* 0x1CA2E */ s8 defense[3];
-    /* 0x1CA32 */ u16 health[3];
-    /* 0x1CA38 */ s16 buttonIndex;
-    /* 0x1CA3A */ s16 confirmButtonIndex; // 0: yes, 1: quit
-    /* 0x1CA3C */ s16 menuMode;
-    /* 0x1CA3E */ s16 configMode;
-    /* 0x1CA40 */ s16 prevConfigMode;
-    /* 0x1CA42 */ s16 nextConfigMode;
-    /* 0x1CA44 */ s16 selectMode;
-    /* 0x1CA46 */ s16 selectedFileIndex;
-    /* 0x1CA48 */ char unk_1CA48[0x2];
-    /* 0x1CA4A */ s16 fileNamesY[3];
-    /* 0x1CA50 */ s16 actionTimer;
-    /* 0x1CA52 */ s16 buttonYOffsets[6];
-    /* 0x1CA5E */ s16 copyDestFileIndex;
-    /* 0x1CA60 */ s16 warningLabel;
-    /* 0x1CA62 */ s16 warningButtonIndex;
-    /* 0x1CA64 */ s16 titleLabel;
-    /* 0x1CA66 */ s16 nextTitleLabel;
-    /* 0x1CA68 */ s16 windowColor[3];
-    /* 0x1CA6E */ s16 titleAlpha[2];
-    /* 0x1CA72 */ s16 windowAlpha;
-    /* 0x1CA74 */ s16 fileButtonAlpha[3];
-    /* 0x1CA7A */ s16 nameBoxAlpha[3];
-    /* 0x1CA80 */ s16 nameAlpha[3];
-    /* 0x1CA86 */ s16 connectorAlpha[3];
-    /* 0x1CA8C */ s16 fileInfoAlpha[3];
-    /* 0x1CA92 */ s16 actionButtonAlpha[2];
-    /* 0x1CA96 */ s16 confirmButtonAlpha[2];
-    /* 0x1CA9A */ s16 optionButtonAlpha;
-    /* 0x1CA9C */ s16 nameEntryBoxAlpha;
-    /* 0x1CA9E */ s16 controlsAlpha;
-    /* 0x1CAA0 */ s16 emptyFileTextAlpha;
-    /* 0x1CAA2 */ s16 highlightColor[4];
-    /* 0x1CAAA */ s16 highlightPulseDir; // 0 fade out, 1 fade in
-    /* 0x1CAAC */ s16 unk_1CAAC; // initialized but never used
-    /* 0x1CAAE */ s16 confirmButtonTexIndices[2];
-    /* 0x1CAB2 */ s16 inputTimerX;
-    /* 0x1CAB4 */ s16 inputTimerY;
-    /* 0x1CAB6 */ s16 stickXDir;
-    /* 0x1CAB8 */ s16 stickYDir;
-    /* 0x1CABA */ s16 stickAdjX;
-    /* 0x1CABC */ s16 stickAdjY;
-    /* 0x1CABE */ s16 nameEntryBoxPosX;
-    /* 0x1CAC0 */ s16 windowPosX;
-    /* 0x1CAC4 */ f32 windowRot;
-    /* 0x1CAC8 */ s16 kbdButton; // only for buttons, not characters
-    /* 0x1CACA */ s16 charPage; // 0: hiragana, 1: katakana, 2: alphabet
-    /* 0x1CACC */ s16 charBgAlpha; // square shape the letter sits in
-    /* 0x1CACE */ s16 charIndex; // 0 - 64, top left to bottom right
-    /* 0x1CAD0 */ s16 kbdX; // (0, 0) is top left character
-    /* 0x1CAD2 */ s16 kbdY;
-    /* 0x1CAD4 */ s16 newFileNameCharCount;
-    /* 0x1CAD6 */ s16 unk_1CAD6[5];
-} FileSelectState; // size = 0x1CAE0
 
 // Macros for `EntranceInfo.field`
 #define ENTRANCE_INFO_CONTINUE_BGM_FLAG (1 << 15)
