+++ src/overlays/gamestates/ovl_file_choose/z_file_choose.c
@@ -1,5 +1,6 @@
 #include "file_select.h"
 #include "terminal.h"
+#include "overlays/gamestates/ovl_select/z_select.h"
 #include "assets/textures/title_static/title_static.h"
 #include "assets/textures/parameter_static/parameter_static.h"
