+++ src/overlays/gamestates/ovl_title/z_title.h
@@ -1,0 +1,27 @@
+#ifndef Z_TITLE_H
+#define Z_TITLE_H
+
+#include "global.h"
+#include "z64game.h"
+#include "z64view.h"
+
+typedef struct {
+    /* 0x0000 */ GameState state;
+    /* 0x00A4 */ u8* staticSegment;
+    /* 0x00A8 */ View view;
+    /* 0x01D0 */ SramContext sramCtx;
+    /* 0x01D4 */ u16 unk_1D4; // not used in mq dbg (some sort of timer that doesn't seem to affect anything)
+    /* 0x01D6 */ s16 coverAlpha;
+    /* 0x01D8 */ s16 addAlpha; // not used in mq dbg
+    /* 0x01DA */ u16 visibleDuration; // not used in mq dbg
+    /* 0x01DC */ s16 ult;
+    /* 0x01DE */ s16 uls;
+    /* 0x01E0 */ char unk_1E0[0x01];
+    /* 0x01E1 */ u8 exit;
+    /* 0x01E2 */ char unk_1E2[0x06];
+} ConsoleLogoState; // size = 0x1E8
+
+void ConsoleLogo_Init(GameState* thisx);
+void ConsoleLogo_Destroy(GameState* thisx);
+
+#endif
