+++ src/overlays/misc/ovl_kaleido_scope/z_kaleido_scope_PAL.c
@@ -1,4 +1,5 @@
 #include "z_kaleido_scope.h"
+#include "overlays/gamestates/ovl_opening/z_opening.h"
 #include "assets/textures/icon_item_static/icon_item_static.h"
 #include "assets/textures/icon_item_24_static/icon_item_24_static.h"
 #include "assets/textures/icon_item_nes_static/icon_item_nes_static.h"
